Fix more compile warnings from jtarchie's upload form code.

+++ ext/typhoeus/typhoeus_form.c
@@ -15,7 +15,7 @@ static VALUE formadd_file(VALUE self,
   CurlForm *curl_form;
   Data_Get_Struct(self, CurlForm, curl_form);
 
-  return INT2NUM((int)curl_formadd(&curl_form->first, &curl_form->last,
+  return INT2NUM((long) curl_formadd(&curl_form->first, &curl_form->last,
     CURLFORM_COPYNAME, RSTRING(name)->ptr,
     CURLFORM_NAMELENGTH, (long)RSTRING(name)->len,
     CURLFORM_FILENAME, RSTRING(filename)->ptr,
@@ -23,14 +23,13 @@ static VALUE formadd_file(VALUE self,
     CURLFORM_FILE, RSTRING(path)->ptr,
     CURLFORM_END
   ));
-
 }
 
 static VALUE formadd_param(VALUE self, VALUE name, VALUE value) {
   CurlForm *curl_form;
   Data_Get_Struct(self, CurlForm, curl_form);
 
-  return INT2NUM((int)curl_formadd(&curl_form->first, &curl_form->last,
+  return INT2NUM((long)curl_formadd(&curl_form->first, &curl_form->last,
     CURLFORM_COPYNAME, RSTRING(name)->ptr,
     CURLFORM_NAMELENGTH, (long)RSTRING(name)->len,
     CURLFORM_COPYCONTENTS, RSTRING(value)->ptr,
